+++ packages/core/client/src/schema-component/antd/table-v2/Table.tsx
@@ -1,11 +1,13 @@
 import { MenuOutlined } from '@ant-design/icons';
+import { SortableContext, useSortable } from '@dnd-kit/sortable';
 import { css } from '@emotion/css';
-import { ArrayField } from '@formily/core';
+import { ArrayField, Field } from '@formily/core';
 import { ISchema, observer, RecursionField, Schema, useField, useFieldSchema } from '@formily/react';
+import { reaction } from '@formily/reactive';
+import { useMemoizedFn } from 'ahooks';
 import { Table as AntdTable, TableColumnProps } from 'antd';
 import { default as classNames, default as cls } from 'classnames';
-import React from 'react';
-import ReactDragListView from 'react-drag-listview';
+import React, { useCallback, useEffect, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { DndContext } from '../..';
 import { RecordIndexProvider, RecordProvider, useSchemaInitializer } from '../../../';
@@ -16,7 +18,7 @@ const isColumnComponent = (schema: Schema) => {
 
 const isCollectionFieldComponent = (schema: ISchema) => {
   return schema['x-component'] === 'CollectionField';
-}
+};
 
 const useTableColumns = () => {
   const start = Date.now();
@@ -30,12 +32,11 @@ const useTableColumns = () => {
       }
     }, [])
     .map((s: Schema) => {
-      const collectionFields = s
-        .reduceProperties((buf, s) => {
-          if (isCollectionFieldComponent(s)) {
-            return buf.concat([s]);
-          }
-        }, []);
+      const collectionFields = s.reduceProperties((buf, s) => {
+        if (isCollectionFieldComponent(s)) {
+          return buf.concat([s]);
+        }
+      }, []);
       const dataIndex = collectionFields?.length > 0 ? collectionFields[0].name : s.name;
 
       return {
@@ -67,64 +68,42 @@ const useTableColumns = () => {
   });
 };
 
-export const components = {
-  header: {
-    wrapper: (props) => {
-      return (
-        <DndContext>
-          <thead {...props} />
-        </DndContext>
-      );
-    },
-    cell: (props) => {
-      return (
-        <th
-          {...props}
-          className={cls(
-            props.className,
-            css`
-              max-width: 300px;
-              white-space: nowrap;
-              &:hover .general-schema-designer {
-                display: block;
-              }
-            `,
-          )}
-        />
-      );
-    },
-  },
-  body: {
-    wrapper: (props) => {
-      return (
-        <DndContext>
-          <tbody {...props} />
-        </DndContext>
-      );
-    },
-    row: (props) => {
-      return <tr {...props} />;
-    },
-    cell: (props) => (
-      <td
-        {...props}
-        className={classNames(
-          props.className,
-          css`
-            max-width: 300px;
-            white-space: nowrap;
-            .nb-read-pretty-input-number {
-              text-align: right;
-            }
-          `,
-        )}
-      />
-    ),
-  },
+const topActiveClass = css`
+  & > td {
+    border-top: 2px solid rgba(241, 139, 98, 0.6) !important;
+  }
+`;
+const bottomActiveClass = css`
+  & > td {
+    border-bottom: 2px solid rgba(241, 139, 98, 0.6) !important;
+  }
+`;
+
+const SortableRow = (props) => {
+  const id = props['data-row-key']?.toString();
+  const { setNodeRef, isOver, active, over } = useSortable({
+    id,
+  });
+
+  const className =
+    (active?.data.current?.sortable.index ?? -1) > (over?.data.current?.sortable.index ?? -1)
+      ? topActiveClass
+      : bottomActiveClass;
+
+  return (
+    <tr
+      ref={active?.id !== id ? setNodeRef : null}
+      {...props}
+      className={classNames({ [className]: active && isOver })}
+    />
+  );
 };
 
-const SortHandle = () => {
-  return <MenuOutlined className={'drag-handle'} style={{ cursor: 'grab' }} />;
+const SortHandle = (props) => {
+  const { listeners } = useSortable({
+    id: props.id,
+  });
+  return <MenuOutlined {...listeners} style={{ cursor: 'grab' }} />;
 };
 
 const TableIndex = (props) => {
@@ -152,6 +131,26 @@ const usePaginationProps = (pagination1, pagination2) => {
   };
 };
 
+const useValidator = (validator: (value: any) => string) => {
+  const field = useField<Field>();
+  useEffect(() => {
+    const dispose = reaction(
+      () => field.value,
+      (value) => {
+        const message = validator(value);
+        field.setFeedback({
+          type: 'error',
+          code: 'ValidateError',
+          messages: message ? [message] : [],
+        });
+      },
+    );
+    return () => {
+      dispose();
+    };
+  }, []);
+};
+
 export const Table: any = observer((props: any) => {
   const field = useField<ArrayField>();
   const columns = useTableColumns();
@@ -160,13 +159,108 @@ export const Table: any = observer((props: any) => {
   const {
     dragSort = false,
     showIndex = true,
-    onRowDragEnd,
     onRowSelectionChange,
     onChange: onTableChange,
     rowSelection,
+    rowKey,
+    required,
     ...others
   } = { ...others1, ...others2 } as any;
+  const onRowDragEnd = useMemoizedFn(others.onRowDragEnd || (() => {}));
   const paginationProps = usePaginationProps(pagination1, pagination2);
+  const requiredValidator = field.required || required;
+  useEffect(() => {
+    field.setValidator((value) => {
+      if (requiredValidator) {
+        return Array.isArray(value) && value.length > 0 ? null : 'The field value is required';
+      }
+      return;
+    });
+  }, [requiredValidator]);
+  const components = useMemo(() => {
+    return {
+      header: {
+        wrapper: (props) => {
+          return (
+            <DndContext>
+              <thead {...props} />
+            </DndContext>
+          );
+        },
+        cell: (props) => {
+          return (
+            <th
+              {...props}
+              className={cls(
+                props.className,
+                css`
+                  max-width: 300px;
+                  white-space: nowrap;
+                  &:hover .general-schema-designer {
+                    display: block;
+                  }
+                `,
+              )}
+            />
+          );
+        },
+      },
+      body: {
+        wrapper: (props) => {
+          return (
+            <DndContext
+              onDragEnd={(e) => {
+                if (!e.active || !e.over) {
+                  console.warn('move cancel');
+                  return;
+                }
+
+                const fromIndex = e.active?.data.current?.sortable?.index;
+                const toIndex = e.over?.data.current?.sortable?.index;
+                const from = field.value[fromIndex];
+                const to = field.value[toIndex];
+                field.move(fromIndex, toIndex);
+                onRowDragEnd({ fromIndex, toIndex, from, to });
+              }}
+            >
+              <tbody {...props} />
+            </DndContext>
+          );
+        },
+        row: (props) => {
+          return <SortableRow {...props}></SortableRow>;
+        },
+        cell: (props) => (
+          <td
+            {...props}
+            className={classNames(
+              props.className,
+              css`
+                max-width: 300px;
+                white-space: nowrap;
+                .nb-read-pretty-input-number {
+                  text-align: right;
+                }
+              `,
+            )}
+          />
+        ),
+      },
+    };
+  }, [field, onRowDragEnd, dragSort]);
+
+  const defaultRowKey = (record: any) => {
+    return field.value?.indexOf?.(record);
+  };
+
+  const getRowKey = (record: any) => {
+    if (typeof rowKey === 'string') {
+      return record[rowKey]?.toString();
+    } else {
+      return (rowKey ?? defaultRowKey)(record)?.toString();
+    }
+  };
+
   const restProps = {
     rowSelection: rowSelection
       ? {
@@ -226,7 +320,7 @@ export const Table: any = observer((props: any) => {
                     `,
                   )}
                 >
-                  {dragSort && <SortHandle />}
+                  {dragSort && <SortHandle id={getRowKey(record)} />}
                   {showIndex && <TableIndex index={index} />}
                 </div>
                 <div
@@ -252,9 +346,21 @@ export const Table: any = observer((props: any) => {
         }
       : undefined,
   };
-  const defaultRowKey = (record: any) => {
-    return field.value?.indexOf?.(record);
-  };
+
+  const SortableWrapper = useCallback<React.FC>(
+    ({ children }) => {
+      return dragSort
+        ? React.createElement(SortableContext, {
+            items: field.value.map(getRowKey),
+            children: children,
+          })
+        : React.createElement(React.Fragment, {
+            children,
+          });
+    },
+    [field, dragSort],
+  );
+
   return (
     <div
       className={css`
@@ -264,20 +370,9 @@ export const Table: any = observer((props: any) => {
         }
       `}
     >
-      <ReactDragListView
-        handleSelector={'.drag-handle'}
-        onDragEnd={async (fromIndex, toIndex) => {
-          const from = field.value[fromIndex];
-          const to = field.value[toIndex];
-          field.move(fromIndex, toIndex);
-          onRowDragEnd({ fromIndex, toIndex, from, to });
-        }}
-        lineClassName={css`
-          border-bottom: 2px solid rgba(241, 139, 98, 0.6) !important;
-        `}
-      >
+      <SortableWrapper>
         <AntdTable
-          rowKey={defaultRowKey}
+          rowKey={rowKey ?? defaultRowKey}
           {...others}
           {...restProps}
           pagination={paginationProps}
@@ -290,7 +385,14 @@ export const Table: any = observer((props: any) => {
           columns={columns}
           dataSource={field?.value?.slice?.()}
         />
-      </ReactDragListView>
+      </SortableWrapper>
+      {field.errors.length > 0 && (
+        <div className="ant-formily-item-error-help ant-formily-item-help ant-formily-item-help-enter ant-formily-item-help-enter-active">
+          {field.errors.map((error) => {
+            return error.messages.map((message) => <div>{message}</div>);
+          })}
+        </div>
+      )}
     </div>
   );
 });