+++ packages/plugins/users/src/collections/users.ts
@@ -37,6 +37,7 @@ export default {
         type: 'string',
         title: '{{t("Email")}}',
         'x-component': 'Input',
+        'x-validator': 'email',
         require: true,
       },
     },
@@ -49,29 +50,6 @@ export default {
         type: 'string',
         title: '{{t("Password")}}',
         'x-component': 'Password',
-      },
-    },
-    {
-      interface: 'm2m',
-      type: 'belongsToMany',
-      name: 'roles',
-      target: 'roles',
-      foreignKey: 'userId',
-      otherKey: 'roleName',
-      sourceKey: 'id',
-      targetKey: 'name',
-      through: 'rolesUsers',
-      uiSchema: {
-        type: 'array',
-        title: '{{t("Roles")}}',
-        'x-component': 'RecordPicker',
-        'x-component-props': {
-          multiple: true,
-          fieldNames: {
-            label: 'title',
-            value: 'name',
-          },
-        },
       },
     },
     {