+++ packages/plugins/workflow/src/server/actions/index.ts
@@ -1,5 +1,6 @@
 import * as workflows from './workflows';
 import * as nodes from './nodes';
+import * as jobs from './jobs';
 
 function make(name, mod) {
   return Object.keys(mod).reduce((result, key) => ({
@@ -17,6 +18,7 @@ export default function({ app }) {
     }),
     ...make('flow_nodes', {
       update: nodes.update
-    })
+    }),
+    ...make('jobs', jobs)
   });
 }